move ControlBlock to its own module
